Add base_frame class to apply inheritance
