+++ base_frame.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+# BaseFrame class to inheritance methods
+class BaseFrame
+  def strike?
+    @turns.size == 1 && @turns.sum == 10
+  end
+
+  def add(pin_down)
+    @turns << pin_down
+  end
+
+  def turns_completed?
+    @turns.size == 2
+  end
+end
