+++ game_roll.rb
@@ -1,32 +1,22 @@
 # frozen_string_literal: true
 
+require_relative 'base_frame.rb'
+
 # game roll logic class
-class GameRoll
-    def initialize
-      @turns = []
-    end
-    
-    def strike?
-      @turns.size == 1 && @turns.sum == 10
-    end
-    
-    def completed?
-      @turns.size == 2
-    end
-    
-    def add(pin_down)
-      @turns << pin_down
-    end
-    
-    def first_turn_pins_down
-      @turns.first
-    end
-    
-    def second_turn_pins_down
-      @turns.last
-    end
-    
-    def reset
-      @turns = []
-    end
+class GameRoll < BaseFrame
+  def initialize
+    @turns = []
   end
+
+  def first_turn_pins_down
+    @turns.first
+  end
+
+  def second_turn_pins_down
+    @turns.last
+  end
+
+  def reset
+    @turns = []
+  end
+end
